requirements.txt
typer==0.7.0
rich==12.6.0
pyyaml==6.0
